Updated HutosLib and added Guide

+++ src/main/java/com/vincenthuto/hemomancy/network/particle/PacketSpawnFlaskParticles.java
@@ -3,9 +3,9 @@ package com.vincenthuto.hemomancy.network.particle;
 import java.util.Optional;
 import java.util.function.Supplier;
 
-import com.hutoslib.client.particle.factory.GlowParticleFactory;
-import com.hutoslib.client.particle.util.ParticleColor;
-import com.hutoslib.client.particle.util.ParticleUtils;
+import com.vincenthuto.hutoslib.client.particle.factory.GlowParticleFactory;
+import com.vincenthuto.hutoslib.client.particle.util.ParticleColor;
+import com.vincenthuto.hutoslib.client.particle.util.ParticleUtils;
 
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.network.FriendlyByteBuf;
@@ -64,8 +64,8 @@ public class PacketSpawnFlaskParticles {
 			return;
 		}
 		ClientLevel world = ((ClientLevel) clientLevel.get());
-		world.addParticle(GlowParticleFactory.createData(msg.getColor()), (double) msg.getPos().x, msg.getPos().y + 1,
-				(double) msg.getPos().z, ParticleUtils.inRange(-3, 3) * 0.015f, ParticleUtils.inRange(-3, 3) * 0.015f,
+		world.addParticle(GlowParticleFactory.createData(msg.getColor()), msg.getPos().x, msg.getPos().y + 1,
+				msg.getPos().z, ParticleUtils.inRange(-3, 3) * 0.015f, ParticleUtils.inRange(-3, 3) * 0.015f,
 				ParticleUtils.inRange(-3, 3) * 0.015f);
 
 		ctxSupplier.get().setPacketHandled(true);
